+++ datapilot/visualization/__init__.py
@@ -3,5 +3,6 @@ from .boxplot import box
 from .heatmap import heatmap
 from .scatter import scatter
 from .violin import violin
+from .time_series import plot_time_series
 
-__all__ = ["hist", "box", "heatmap", "scatter", "violin"]
+__all__ = ["hist", "box", "heatmap", "scatter", "violin", "plot_time_series"]